clearing input after creating post
